Fixed double vs float issues in gadget functionalities; fixed r_s rescaling issue in drawDensProfs(); allow user to choose between Powell, TNC and other minimization methods

+++ cosmic_profiles/for_docs/gadget/readgadget.py
@@ -108,7 +108,7 @@ def read_field(snapshot, block, ptype):
         
         if '%s%s'%(prefix,suffix) not in f:
             if Masses[ptype] != 0.0:
-                array = np.ones(Npart[ptype], np.float32)*Masses[ptype]/unit_fac
+                array = np.ones(Npart[ptype], np.float64)*Masses[ptype]/unit_fac
             else:
                 raise Exception('Problem reading the block %s'%block)
         else:
@@ -117,7 +117,7 @@ def read_field(snapshot, block, ptype):
 
         if block=="VEL ":  array *= np.sqrt(head.time) # To make sure we move from UnitVelocity a**0.5 to UnitVelocity
         if block=="POS " and array.dtype==np.float64:
-            array = array.astype(np.float32)
+            array = array.astype(np.float64)
 
         return array
 
@@ -144,16 +144,16 @@ def read_block(snapshot, block, ptype, verbose=False):
 
     if block=="POS ":  
         unit_fac = l_target_over_curr
-        dtype=np.dtype((np.float32,3))
+        dtype=np.dtype((np.float64,3))
     elif block=="MASS":  
         unit_fac = m_target_over_curr
-        dtype=np.float32
+        dtype=np.float64
     elif block=="ID  ":  
         unit_fac = 1
         dtype=read_field(filename, block, ptype[0]).dtype
     elif block=="VEL ":  
         unit_fac = v_target_over_curr
-        dtype=np.dtype((np.float32,3))
+        dtype=np.dtype((np.float64,3))
     else: 
         raise Exception('block not implemented in readgadget!')
